requirements.txt
pylast
python-socketio<6 # To ensure Engine 4 support
python-dotenv